cli/medperf/templates/federated_server_mlcube/{{cookiecutter.project_slug}}/project/requirements.txt
flwr>=1.8.0
PyYAML==6.0
